Factor out core parts of EARCompiler
