+++ ear-compiler/src/main/java/eflang/ear/core/Instruction.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 public class Instruction {
     public enum Type {
