+++ ear-compiler/src/main/java/eflang/ear/operation/Subtract.java
@@ -1,7 +1,7 @@
 package eflang.ear.operation;
 
 import eflang.ear.Argument;
-import eflang.ear.Instruction;
+import eflang.ear.core.Instruction;
 
 import java.util.ArrayList;
 import java.util.List;
